almide-wasm-run to codopsy A: the p3 and p1 shim files split into include! parts, shim parameters bundled into structs, and the host dispatch split by op family
